+++ CHANGES.rst
@@ -1,5 +1,13 @@
 Changelog
 =========
+Changes in 0.2.0
+----------------
+ * Kramer-Kronig relation implementation added to utils
+ * Example for using the KK relation added to documentation
+ * Tests for KK relation
+ * Added pip-requirements file
+ * Fixed issue #16 of spectrum convolution breaking units
+
 Changes in 0.1.2
 ----------------
  * Fixed issue #12 of interpolation breaking units